Make a SlackIntegration component
