+++ frontend/src/pages/Setup/SetupPage.module.scss
@@ -105,10 +105,6 @@
     font-weight: 200;
 }
 
-.alertsPage {
-    padding: 0;
-}
-
 .codeBlockBasic {
     font-family: 'Lucida Console', Monaco, Consolas, 'Courier New', Courier,
         'Lucida Sans Typewriter', 'Lucida Typewriter', monospace !important;
